Added the option to execute user defined test queries after chasing.
https://github.com/michaelbenedikt/pdq/issues/347

+++ reasoning/test/src/uk/ac/ox/cs/pdq/test/reasoning/ReasonTest.java
@@ -40,30 +40,67 @@ public class ReasonTest {
 		}
 		try {
 			Reason r = new Reason(new String[] { "-s", schema.getAbsolutePath(), "-q", query.getAbsolutePath(), "-c",
-					config.getAbsolutePath(), "-o", output.getAbsolutePath(), "-ca" });
-			Assert.assertTrue(r.isCaOnly());
-			Assert.assertFalse(r.isVerbose());
-			File generatedOutput = new File(output, "part.csv");
-			Assert.assertTrue(generatedOutput.exists());
-			Assert.assertTrue(generatedOutput.length() > 1);
-			generatedOutput.delete();
-			File generatedOutput2 = new File(output, "partsupp.csv");
-			Assert.assertTrue(generatedOutput2.exists());
-			generatedOutput2.delete();
-			Assert.assertTrue(output.list().length == 0);
-
-			Reason r2 = new Reason(new String[] { "-s", schema.getAbsolutePath(), "-q", query.getAbsolutePath(), "-c",
 					config.getAbsolutePath(), "-o", output.getAbsolutePath() });
-			Assert.assertFalse(r2.isCaOnly());
-			Assert.assertTrue(output.list().length == 6);
-			for (File child : output.listFiles())
-				child.delete();
-			output.delete();
+			Assert.assertFalse(r.isVerbose());
+			testAndDelete(new File(output, "part.csv"));
+			testAndDelete(new File(output, "partsupp.csv"));
+			testAndDelete(new File(output, "nation.csv"));
+			testAndDelete(new File(output, "region.csv"));
+			testAndDelete(new File(output, "region_nation.csv"));
+			testAndDelete(new File(output, "supplier.csv"));
+
+			Assert.assertTrue(output.list().length == 0);
 		} catch (Exception e) {
 			e.printStackTrace();
 			fail("error " + e.getMessage());
+		} finally {
+			for (File child : output.listFiles())
+				child.delete();
+			output.delete();
 		}
 
+	}
+	@Test
+	public void testUserQueries() {
+		File testFolder = new File("../regression/test/planner/linear/fast/tpch/simple/case_002");
+		File schema = new File(testFolder, "schema.xml");
+		File config = new File(testFolder, "case.properties");
+		File query = new File(testFolder, "query.xml");
+		File output = new File(testFolder, "results//");
+		if (output.exists()) {
+			for (File child : output.listFiles())
+				child.delete();
+		} else {
+			output.mkdirs();
+		}
+		try {
+			Reason r = new Reason(new String[] { "-s", schema.getAbsolutePath(), "-q", query.getAbsolutePath(), "-c",
+					config.getAbsolutePath(), "-o", output.getAbsolutePath(),"-tq", query.getAbsolutePath() });
+			Assert.assertFalse(r.isVerbose());
+			testAndDelete(new File(output, "part.csv"));
+			testAndDelete(new File(output, "partsupp.csv"));
+			testAndDelete(new File(output, "nation.csv"));
+			testAndDelete(new File(output, "region.csv"));
+			testAndDelete(new File(output, "region_nation.csv"));
+			testAndDelete(new File(output, "supplier.csv"));
+			new File(testFolder, "UserQueryResults/query.csv").delete();
+			new File(testFolder, "UserQueryResults").delete();
+			Assert.assertTrue(output.list().length == 0);
+		} catch (Exception e) {
+			e.printStackTrace();
+			fail("error " + e.getMessage());
+		} finally {
+			for (File child : output.listFiles())
+				child.delete();
+			output.delete();
+		}
+
+	}
+
+	private void testAndDelete(File file) {
+		Assert.assertTrue(file.exists());
+		Assert.assertTrue(file.length() > 1);
+		Assert.assertTrue(file.delete());
 	}
 
 }
